Update 2017-09-29-use-rails-naming-conventions-for-dates-and-times.md
> However, on many applications, diving into a `schema.rb` or migration often reveals `something_date` as a field name on a model.

This sentence isn't clear. Can you try and explain in a different way what you mean? Is this `something_date` coming from something the developer has done? And so what is the problem? Are you trying to say that we should be using the default managed timestamps instead of creating our own?

I like your examples!

 >Rails’ own timestamps give us a convention we can reuse and it’s a good idea to suffix times in the same way.

Deleted this because you still haven't got to the 'But why?'. By this point, we know that this is a Rails convention and that you're telling us to suffix it in the same way. So tell us: why?